refactor: RouteModel 패키지 이동
