+++ core/data/auth/src/main/java/com/goalpanzi/mission_mate/core/data/auth/AuthTokenExpirationHandler.kt
@@ -4,7 +4,7 @@ import com.goalpanzi.mission_mate.core.datastore.datasource.AuthDataSource
 import com.goalpanzi.mission_mate.core.datastore.datasource.DefaultDataSource
 import com.goalpanzi.mission_mate.core.datastore.datasource.MissionDataSource
 import com.goalpanzi.mission_mate.core.navigation.NavigationEventHandler
-import com.goalpanzi.mission_mate.core.navigation.RouteModel
+import com.goalpanzi.mission_mate.core.navigation.model.RouteModel
 import com.goalpanzi.mission_mate.core.navigation.di.AuthNavigation
 import com.goalpanzi.mission_mate.core.network.TokenExpirationHandler
 import kotlinx.coroutines.flow.collect
