+++ core/navigation/src/main/java/com/goalpanzi/mission_mate/core/navigation/AuthNavigationEventHandler.kt
@@ -1,5 +1,6 @@
 package com.goalpanzi.mission_mate.core.navigation
 
+import com.goalpanzi.mission_mate.core.navigation.model.RouteModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
