+++ feature/board/src/main/java/com/goalpanzi/mission_mate/feature/board/screen/BoardFinishViewModel.kt
@@ -10,7 +10,7 @@ import com.goalpanzi.mission_mate.core.domain.mission.usecase.CompleteMissionUse
 import com.goalpanzi.mission_mate.core.domain.mission.usecase.GetMissionRankUseCase
 import com.goalpanzi.mission_mate.core.domain.mission.usecase.SetMissionJoinedUseCase
 import com.goalpanzi.mission_mate.core.domain.user.usecase.ProfileUseCase
-import com.goalpanzi.mission_mate.core.navigation.RouteModel.MainTabRoute.MissionRouteModel
+import com.goalpanzi.mission_mate.core.navigation.model.RouteModel.MainTabRoute.MissionRouteModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
