+++ feature/login/src/main/java/com/goalpanzi/mission_mate/feature/login/LoginNavigation.kt
@@ -3,7 +3,7 @@ package com.goalpanzi.mission_mate.feature.login
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.goalpanzi.mission_mate.core.navigation.RouteModel
+import com.goalpanzi.mission_mate.core.navigation.model.RouteModel
 
 fun NavController.navigateToLogin() {
     this.navigate(RouteModel.Login) {
